Move windows CLI handling into separate file and conditionally import for mac compatibility.
